requirements.txt
# Python 3.11
Flask==3.1.2
web3==7.14.0
py-solc-x==2.0.5
requests==2.32.5
supabase==2.27.1